dev-requirements.txt
-r requirements.txt

coala-bears==0.8.4
django-debug-toolbar==1.5
django-nose==1.4.2
